+++ api/names.go
@@ -1,4 +1,4 @@
-package utils
+package api
 
 import "regexp"
